Refactor AI chat UI into modular widgets
Split chat page UI into dedicated widgets for composer, settings bottom sheet, empty state, message bubble, and typing indicator. Removed unused imports and moved feature flag for update to false. Improves maintainability and readability of the chat feature.
