+++ lib/features/ai_chat/widgets/typing_indicator.dart
@@ -1,0 +1,84 @@
+import 'package:flutter/material.dart';
+
+class TypingIndicator extends StatelessWidget {
+  const TypingIndicator({super.key, required this.colorScheme});
+
+  final ColorScheme colorScheme;
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      margin: const EdgeInsets.only(bottom: 8),
+      padding: const EdgeInsets.symmetric(horizontal: 16),
+      child: Row(
+        children: [
+          Container(
+            padding: const EdgeInsets.all(8),
+            decoration: BoxDecoration(
+              color: colorScheme.primary.withValues(alpha: 0.1),
+              borderRadius: BorderRadius.circular(20),
+            ),
+            child: Icon(
+              Icons.smart_toy_outlined,
+              color: colorScheme.primary,
+              size: 16,
+            ),
+          ),
+          const SizedBox(width: 8),
+          Container(
+            padding: const EdgeInsets.all(16),
+            decoration: BoxDecoration(
+              color: colorScheme.surface,
+              borderRadius: BorderRadius.circular(16).copyWith(
+                bottomLeft: const Radius.circular(4),
+              ),
+              border: Border.all(
+                color: colorScheme.outline.withValues(alpha: 0.2),
+                width: 1,
+              ),
+            ),
+            child: Row(
+              mainAxisSize: MainAxisSize.min,
+              children: const [
+                _TypingDot(index: 0),
+                SizedBox(width: 4),
+                _TypingDot(index: 1),
+                SizedBox(width: 4),
+                _TypingDot(index: 2),
+              ],
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
+
+class _TypingDot extends StatelessWidget {
+  const _TypingDot({required this.index});
+  final int index;
+
+  @override
+  Widget build(BuildContext context) {
+    final ColorScheme colorScheme = Theme.of(context).colorScheme;
+    return TweenAnimationBuilder<double>(
+      duration: Duration(milliseconds: 600 + (index * 200)),
+      tween: Tween(begin: 0.0, end: 1.0),
+      curve: Curves.easeInOutSine,
+      builder: (context, value, child) {
+        return Transform.scale(
+          scale: 0.5 + (0.5 * value),
+          child: child!,
+        );
+      },
+      child: Container(
+        width: 8,
+        height: 8,
+        decoration: BoxDecoration(
+          color: colorScheme.primary.withValues(alpha: 0.6),
+          shape: BoxShape.circle,
+        ),
+      ),
+    );
+  }
+}
